Revert "Resolved merge conflict in generator-template.md"
This reverts commit 9040e986e9be2c426e1f02b56c3b7f26ebb42c5e, reversing
changes made to 6e8a89336b4fc5f380236ae37e969218ce4ea459.
