+++ .github/workflows/pr-testing-with-test-project.yml
@@ -9,15 +9,14 @@ jobs:
   changes:
     runs-on: ubuntu-latest
     outputs:
-      should_test: ${{ steps.filter.outputs.modified_files !='[]' }}
+      should_test: ${{ steps.filter.outputs.changes }}
     steps:
       - uses: actions/checkout@f43a0e5ff2bd294095638e18286ca9a3d1956744 # v3 https://github.com/actions/checkout/releases/tag/v3
-      - uses: rohansingh/paths-filter@084ca29929a4e7e708a4aa8b790347facfd830c9 # fork with predicate-quantifier fix
-        #TODO check if we can switch back to https://github.com/dorny/paths-filter in a month or 2 (today 17.03.2025)
+      - uses: dorny/paths-filter@de90cc6fb38fc0963ad72b210f1f284cd68cea36 # v3.0.2 https://github.com/dorny/paths-filter/releases/tag/v3.0.2
+        id: filter
         with:
-          predicate-quantifier: 'every'
           filters: |
-            modified_files:
+            changes:
               - 'package.json'
               - 'package-lock.json'
               - 'apps/**'
@@ -25,7 +24,6 @@ jobs:
               - '.github/workflows/**'
               - '!**/*.md'
               - '!docs/**'
-              - '!.all-contributorsrc'
 
   test:
     needs: changes
