+++ .gitignore
@@ -16,4 +16,3 @@ coverage
 
 /.idea
 temp
-__pycache__
