+++ .sonarcloud.properties
@@ -1,3 +1,2 @@
 #we need to explicitly exclude them as some are commit to the repo
-sonar.exclusions=**/test/**/*,**/*__transpiled/**/*.js,packages/templates/clients/**/*
-#TODO remove packages/templates/clients/**/* when we are more advanced with templates development and the structure of templates is finalized
+sonar.exclusions=**/test/**/*,**/*__transpiled/**/*.js,packages/templates/clients/js/websocket/example.js,
