--- packages/templates/clients/js/websocket/components/Constructor.js
@@ -1,17 +1,0 @@
-import { Text } from '@asyncapi/generator-react-sdk';
-
-export function Constructor({ serverName }) {
-  return (
-    <Text indent={2}>
-      {
-        `constructor() {
-  this.url = '${serverName}';
-  this.websocket = null;
-  this.messageHandlers = [];
-  this.errorHandlers = [];
-}
-`
-      }
-    </Text>
-  );
-}
