--- packages/templates/clients/js/websocket/components/ModuleExport.js
@@ -1,11 +1,0 @@
-import { Text } from '@asyncapi/generator-react-sdk';
-
-export function ModuleExport({ clientName }) {
-  return (
-    <Text>
-      {
-        `module.exports = ${clientName};`
-      }
-    </Text>
-  );
-}
