--- packages/templates/clients/python/websocket/components/ClientClass.js
@@ -1,29 +1,0 @@
-import { Text } from '@asyncapi/generator-react-sdk';
-import { Constructor } from './Constructor';
-import { Connect } from './Connect';
-import { RegisterMessageHandler } from './RegisterMessageHandler';
-import { RegisterErrorHandler } from './RegisterErrorHandler';
-import { HandleMessage } from './HandleMessage';
-import { SendEchoMessage } from './SendEchoMessage';
-import { CloseConnection } from './CloseConnection';
-import { RegisterOutgoingProcessor } from './RegisterOutgoingProcessor';
-import { HandleError } from './HandleError';
-
-export function ClientClass({ clientName, serverUrl, title }) {
-  return (
-    <Text>
-      <Text newLines={2}>
-        {`class ${clientName}:`}
-      </Text>
-      <Constructor serverUrl={serverUrl} />
-      <Connect title={title} />
-      <RegisterMessageHandler />
-      <RegisterErrorHandler />
-      <RegisterOutgoingProcessor />
-      <HandleMessage />
-      <HandleError />
-      <SendEchoMessage />
-      <CloseConnection />
-    </Text>
-  );
-}
