--- packages/templates/clients/python/websocket/components/CloseConnection.js
@@ -1,16 +1,0 @@
-import { Text } from '@asyncapi/generator-react-sdk';
-
-export function CloseConnection() {
-  return (
-    <Text indent={2}>
-      {
-        `def close(self):
-    """Cleanly close the WebSocket connection."""
-    self._stop_event.set()
-    if self.ws_app:
-        self.ws_app.close()
-        print("WebSocket connection closed.")`
-      }
-    </Text>
-  );
-}
