--- packages/templates/clients/python/websocket/components/Requires.js
@@ -1,15 +1,0 @@
-import { Text } from '@asyncapi/generator-react-sdk';
-
-export function Requires() {
-  return (
-    <Text>
-      <Text>
-        {`import json
-import certifi
-import threading
-import websocket`}
-      </Text>
-    </Text>
-  );
-}
-
