--- packages/templates/clients/python/websocket/package.json
@@ -1,67 +1,0 @@
-{
-  "name": "@asyncapi/template-python-websocket-client",
-  "version": "0.0.1",
-  "description": "This is a template generating Python websocket client",
-  "scripts": {
-    "test": "npm run test:cleanup && jest  --coverage",
-    "test:update": "jest --updateSnapshot",
-    "test:cleanup": "rimraf \"tests/temp\""
-  },
-  "author": "Lukasz Gornicki <lpgornicki@gmail.com>",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@asyncapi/generator-react-sdk": "^1.1.2",
-    "@asyncapi/generator-helpers": "*",
-    "rimraf": "^3.0.2"
-  },
-  "devDependencies": {
-    "@asyncapi/parser": "^3.0.14",
-    "@babel/cli": "^7.25.9",
-    "@babel/core": "^7.26.0",
-    "@babel/preset-env": "^7.26.0",
-    "@babel/preset-react": "^7.25.9",
-    "jest-esm-transformer": "^1.0.0",
-    "@asyncapi/generator": "*"
-  },
-  "jest": {
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "jsx"
-    ],
-    "transform": {
-      "^.+\\.jsx?$": "babel-jest"
-    },
-    "moduleNameMapper": {
-      "^nimma/legacy$": "<rootDir>/../../../../../node_modules/nimma/dist/legacy/cjs/index.js",
-      "^nimma/(.*)": "<rootDir>/../../../../../node_modules/nimma/dist/cjs/$1"
-    }
-  },
-  "babel": {
-    "presets": [
-      "@babel/preset-env",
-      [
-        "@babel/preset-react",
-        {
-          "runtime": "automatic"
-        }
-      ]
-    ]
-  },
-  "generator": {
-    "renderer": "react",
-    "apiVersion": "v3",
-    "generator": ">=1.3.0 <3.0.0",
-    "parameters": {
-      "server": {
-        "description": "The name of the server described in AsyncAPI document",
-        "required": true
-      },
-      "clientFileName": {
-        "description": "The name of the generated client file",
-        "required": false,
-        "default": "client.py"
-      }
-    }
-  }
-}
