--- packages/templates/clients/python/websocket/template/client.py.js
@@ -1,23 +1,0 @@
-import { File } from '@asyncapi/generator-react-sdk';
-import { getClientName, getServerUrl } from '@asyncapi/generator-helpers';
-import { FileHeaderInfo } from '../components/FileHeaderInfo';
-import { Requires } from '../components/Requires';
-import { ClientClass } from '../components/ClientClass';
-
-export default function ({ asyncapi, params }) {
-  const server = asyncapi.servers().get(params.server);
-  const info = asyncapi.info();
-  const title = info.title();
-  const clientName = getClientName(info);
-  const serverUrl = getServerUrl(server);
-  return (
-    <File name={params.clientFileName}>
-      <FileHeaderInfo
-        info={info}
-        server={server}
-      />
-      <Requires />
-      <ClientClass clientName={clientName} serverUrl={serverUrl} title={title} />
-    </File>
-  );
-}
