--- packages/templates/clients/python/websocket/test/__fixtures__/asyncapi-hoppscotch-echo.yml
@@ -1,61 +1,0 @@
-asyncapi: 3.0.0
-info:
-  title: Hoppscotch Echo WebSocket
-  version: 1.0.0
-  description: >
-    Undestand how to use Hoppscotch Echo WebSocket as a client.
-    Hoppscotch Echo WebSocket server echoes back any messages sent to it. You can use this to test WebSocket connections and message flows.
-
-servers:
-  echoServer:
-    host: echo-websocket.hoppscotch.io
-    protocol: wss
-
-channels:
-  echo:
-    description: >
-      The single channel where messages are sent to and echoed back. Echo server also regularly drops a timestampl to that channel.
-    address: /
-    messages:
-      echo:
-        $ref: '#/components/messages/echo'
-      timestamp:
-        $ref: '#/components/messages/timestamp'
-
-operations:
-  handleTimeStampMessage:
-    action: receive
-    channel: 
-      $ref: '#/channels/echo'
-    summary: Receive the timestamp message sent from server every second.
-    messages:
-      - $ref: '#/channels/echo/messages/timestamp'
-  # sendEchoMessage:
-  #   action: send
-  #   channel: 
-  #     $ref: '#/channels/echo'
-  #   summary: Send a message to the echo server.
-  #   messages:
-  #     - $ref: '#/channels/echo/messages/echo'
-
-components:
-  messages:
-    timestamp:
-      summary: A message sent from server with current timestamp.
-      payload: 
-        $ref: '#/components/schemas/currentTime'
-    echo:
-      summary: A message exchanged with the echo server.
-      payload: {}
-      examples:
-        - name: string
-          payload: test
-        - name: object
-          payload: 
-            test: test text
-  
-  schemas:
-    currentTime:
-      type: string
-      description: A timestamp with GMT timezone.
-      example: 11:13:24 GMT+0000 (Coordinated Universal Time)
