--- packages/templates/clients/python/websocket/test/__fixtures__/asyncapi-postman-echo.yml
@@ -1,44 +1,0 @@
-asyncapi: 3.0.0
-
-info:
-  title: Postman Echo WebSocket Client
-  version: 1.0.0
-  description: Understand how to use the Postman Echo WebSocket as a client.
-    The Postman Echo WebSocket server echoes back any messages sent to it.
-
-servers:
-  echoServer:
-    host: ws.postman-echo.com
-    pathname: /raw
-    protocol: wss
-
-channels:
-  echo:
-    description: The main channel where messages are sent and echoed back by the Postman Echo WebSocket server.
-    address: /
-    messages:
-      echo:
-        $ref: '#/components/messages/echo'
-
-operations:
-  sendEchoMessage:
-    summary: Send a message to the Postman Echo server.
-    action: send
-    channel: 
-      $ref: '#/channels/echo'
-
-    reply:
-      channel: 
-        $ref: '#/channels/echo'
-
-components:
-  messages:
-    echo:
-      summary: A message exchanged with the echo server.
-      payload: {}
-      examples:
-        - name: string
-          payload: test
-        - name: object
-          payload: 
-            test: test text
